Rename transport interface files
