+++ FreeRTOS-Plus/Demo/FreeRTOS-IoT-Libraries-LTS-Beta2/mqtt/mqtt_plain_text/DemoTasks/PlaintextMQTTExample.c
@@ -58,7 +58,7 @@
 #include "mqtt.h"
 
 /* Transport interface include. */
-#include "transport_interface_freertos.h"
+#include "plaintext_freertos.h"
 
 /*-----------------------------------------------------------*/
